Usando número com virgula

+++ app/src/main/java/Entitys/User.kt
@@ -4,6 +4,6 @@ data class User(
     var email: String,
     var number: String,
     var name: String,
-    var totalMoney: Int = 0,
+    var totalMoney: Double = 0.00,
 
 )
